split test_xarray.py into two files
